backend/requirements.txt
pandas
kafka-python-ng
scikit-learn
joblib